Add files via upload
Added new auto update function. Firmware can be hosted on a webserver. .Bin file needs to be named AtomicSmart.bin and a second file labelled AtomicSmart.version with a simple one line Version number in its body. If auto update enabled then every 60 seconds the update server will be checked for a new version which will be downloaded and installed on the module which will then reboot.